Added install_commands.sh and checks for running iproxy
Added install_commands.sh to list commands that need to be run before Dolosoft can be used.

+++ Dolosoft/AMConnectionHandler.h
@@ -6,6 +6,7 @@
 //  Copyright © 2018 Ander Moran. All rights reserved.
 //
 #import <Foundation/Foundation.h>
+#import <Cocoa/Cocoa.h>
 #import <NMSSH/NMSSH.h>
 
 @interface AMConnectionHandler : NSObject {
@@ -14,5 +15,5 @@
 @property(retain,nonatomic) NMSSHSession *session;
 - (instancetype)initWithHost:(NSString *)host port:(NSInteger)port username:(NSString *)username password:(NSString *)password;
 - (void)initializeProxy;
-- (NMSSHSession *)initializeSessionWithHost:(NSString *)host port:(int)port username:(NSString *)username password:(NSString *)password;
+- (NMSSHSession *)initializeSessionWithHost:(NSString *)host port:(NSInteger)port username:(NSString *)username password:(NSString *)password;
 @end
